Move github module into src/api/
