+++ owners/bootstrap.js
@@ -35,7 +35,7 @@ function bootstrap(logger) {
     require('dotenv').config();
 
     const Octokit = require('@octokit/rest');
-    const {GitHub} = require('./src/github');
+    const {GitHub} = require('./src/api/github');
     const LocalRepository = require('./src/repo/local_repo');
     const {OwnersBot} = require('./src/owners_bot');
