+++ owners/test/owner.test.js
@@ -21,7 +21,7 @@ const {
   WildcardOwner,
   OWNER_MODIFIER,
 } = require('../src/owner');
-const {Team} = require('../src/github');
+const {Team} = require('../src/api/github');
 
 describe('owner base class', () => {
   const owner = new Owner();
